BUG:NS3:lr-wpan-phy: fix CCA m_ccaPeakPower error
//m_ccaPeakPower = 0.0;
m_ccaPeakPower = LrWpanSpectrumValueHelper::TotalAvgPower (m_signal->GetSignalPsd (), m_phyPIBAttributes.phyCurrentChannel);

The initial value of ccaPeakPower should not be 0. This bug is found by
spectrum analyse. When there is a packet that has already in air
transmitting, and finishs before CCA end. If without initial
m_ccaPeakPower with current energy level, The CCA will make mistake that
the PeakPower is 0